Move cansas tests from utest_abs_reader.py to utest_cansas.py and move XML test files into sub-directory.
