--- test/sasdataloader/test/cansas1d_badunits.xml
@@ -1,141 +1,0 @@
-<?xml version="1.0"?>
-<?xml-stylesheet type="text/xsl" href="example.xsl" ?>
-<SASroot version="1.0"
-    xmlns="cansas1d/1.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="cansas1d/1.0 http://svn.smallangles.net/svn/canSAS/1dwg/trunk/cansas1d.xsd"
-    >
-	<SASentry>
-		<Title>Test title</Title>
-		<Run name='run name'>1234</Run>
-		<SASdata>
-			<Idata>
-				<Q unit="1/nm">0.002</Q>
-				<I unit="1/nm">1</I>
-				<Idev unit="1/nm">.003</Idev>
-				<Qdev unit="1/nm">0.001</Qdev>
-				<Qmean unit="1/A"><!-- Qmean is optional --></Qmean>
-				<Shadowfactor><!-- Shadowfactor is optional --></Shadowfactor>
-			</Idata>
-			<Idata>
-				<Q unit="1/nm">0.003</Q>
-				<I unit="1/nm">1.001</I>
-				<Idev unit="1/nm">.004</Idev>
-				<Qdev unit="1/nm">0.002</Qdev>
-				<Qmean unit="1/A"><!-- Qmean is optional --></Qmean>
-				<Shadowfactor><!-- Shadowfactor is optional --></Shadowfactor>
-			</Idata>
-		</SASdata>
-		<SASsample name='my sample'>
-			<ID>SI600-new-long</ID>
-			<thickness unit="mff">0.00103</thickness>
-			<transmission>0.327</transmission>
-			<temperature unit="C">0.0000</temperature>
-			<position>
-				<x unit="nm">10000000.00</x>
-				<y unit="mm">0.00</y>
-			</position>
-			<orientation>
-				<roll unit="rad">0.39269908<!-- was: sample_orientation --></roll>
-				<pitch unit="rad">3.4906585e-4<!-- was: sample_offset_angle --></pitch>
-			</orientation>
-			<details>
-				<!-- was: sample_prep -->
-				http://chemtools.chem.soton.ac.uk/projects/blog/blogs.php/bit_id/2720
-			</details>
-			<details>
-				Some text here
-			</details>
-			
-		</SASsample>
-		<SASinstrument>
-			<name>canSAS instrument</name>
-			<SASsource name="source name">
-				<radiation>neutron</radiation>
-				<beam_size name='bm'>
-					<x unit="m">.01200</x>
-					<y unit="micrometer">13000.00</y>
-				</beam_size>
-				<beam_shape>disc</beam_shape>
-				<wavelength unit="nm">0.600</wavelength>
-				<wavelength_min unit="A">02.2</wavelength_min>
-				<wavelength_max unit="A">10.0</wavelength_max>
-				<wavelength_spread unit="percent">
-					14.3
-				</wavelength_spread>
-			</SASsource>
-			<SAScollimation name="test coll name">
-				<length unit='m'> 0.123</length>
-				<aperture name="source" type="radius">
-					<size>
-						<x unit="micron">50000</x>
-					</size>
-					<distance unit="cm">1100.000<!-- was: distance_coll --></distance>
-				</aperture>
-				<aperture name="sample" type="radius">
-					<size name='size name'>
-						<x unit="cm">.1</x>
-					</size>
-				</aperture>
-			</SAScollimation>
-			<SASdetector>
-				<name>fictional hybrid</name>
-				<SDD unit="m">
-					<!-- sample-to-detector distance -->
-					4.150
-				</SDD>
-				<offset>
-                    <x unit="nm">1000000</x>
-                    <y unit="micron">2000</y>
-                </offset>
-                <orientation>
-					<roll unit="radian">.0174533</roll>
-					<pitch unit="rad">0.00</pitch>
-					<yaw unit="radian">0.00</yaw>
-				</orientation>
-				<beam_center>
-					<x unit="m">.32264</x>
-					<y unit="m">.32768</y>
-				</beam_center>
-				<pixel_size>
-					<x unit="cm">.500</x>
-					<y unit="cm">.500</y>
-				</pixel_size>
-			</SASdetector>
-		</SASinstrument>
-		<SASprocess>
-			<name>spol</name>
-			<date>04-Sep-2007 18:35:02</date>
-			<term name="radialstep" unit="mm">10.000</term>
-			<term name="sector_width" unit="rad">4.1416</term>
-			<term name="sector_orient" unit="degree">0.0</term>
-			<term name="MASK_file">USER:MASK.COM</term>
-			<SASprocessnote>
-				AvA1 0.0000E+00 AsA2 1.0000E+00 XvA3 1.0526E+03 XsA4
-				5.2200E-02 XfA5 0.0000E+00
-			</SASprocessnote>
-			<SASprocessnote>
-				S... 13597 0 2.26E+02 2A 5mM 0%D2O Sbak 13594 0 1.13E+02
-				H2O Buffer
-			</SASprocessnote>
-			<SASprocessnote>V... 13552 3 1.00E+00 H2O5m</SASprocessnote>
-		</SASprocess>
-		<SASprocess>
-			<name>NCNR-IGOR</name>
-			<date>03-SEP-2006 11:42:47</date>
-			<description />
-			<term name="average_type">Circular</term>
-			<term name="SAM_file">SEP06064.SA3_AJJ_L205</term>
-			<term name="BKD_file">SEP06064.SA3_AJJ_L205</term>
-			<term name="EMP_file">SEP06064.SA3_AJJ_L205</term>
-			<term name="DIV_file">SEP06064.SA3_AJJ_L205</term>
-			<term name="MASK_file">SEP06064.SA3_AJJ_L205</term>
-			<term name="ABS:TSTAND">1</term>
-			<term name="ABS:DSTAND" unit="mm">1</term>
-			<term name="ABS:IZERO">230.09</term>
-			<term name="ABS:XSECT" unit="mm">1</term>
-			<SASprocessnote/>
-		</SASprocess>
-		<SASnote />
-	</SASentry>
-</SASroot>
